refactor: update component imports to use tambo paths
