+++ cli/src/registry/message/message.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { createMarkdownComponents } from "@/components/ui/markdown-components";
+import { createMarkdownComponents } from "@/components/tambo/markdown-components";
 import { cn } from "@/lib/utils";
 import type { TamboThreadMessage } from "@tambo-ai/react";
 import { useTambo } from "@tambo-ai/react";
